feat(core): separated `choices` from `Dialogue`
I encountered an issue when trying to categorize action scripts as dialogues. By action script, I mean something along the lines of "Alex opened the door."

Technically, they can't be referred to as dialogues so I had to think of a better way of handling the sequence within a part.

Changes:
- Added a generic trait called `ProgressiveElement<T>`. This would describe part elements so that they always return the desired `Progressive` output.
- Changed `Part` property `dialogues` to `quotes`
- Updated all respective functions related to `quotes`
- Added a `Selection` struct and its respective builder. Selections are the new progressive element that will hold choices.
- Placed all progressive elements under the parts module.